backend/requirements.txt
# Flask and API
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.23

# Google Calendar
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# AI and MCP
anthropic==0.39.0
langchain==0.1.0
langchain-anthropic==0.1.0

# QR Code generation
qrcode==7.4.2
# Pillow==10.4.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3